+++ src/pages/PagesIndex.js
@@ -2,6 +2,6 @@ export { default as Home } from "./Home";
 export { default as Login } from "./Login";
 export { default as Signup } from "./Signup";
 //
-export { default as Reservation } from "./Reservation";
+export { default as Reservation } from "./Reservation/Reservation";
 export { default as Intro } from "./Introduction";
-export { default as Profile } from "./Profile";
+export { default as Profile } from "./Profile/Profile";
