+++ src/pages/Reservation/reservation.css
@@ -1,0 +1,66 @@
+.reservation {
+  position: relative;
+}
+.reservationBackgroundImg {
+  position: relative;
+  width: 100%;
+  height: 100vh;
+  object-fit: cover;
+}
+.reserveInfoWrapper {
+  position: absolute;
+  top: 0;
+  left: 0;
+  height: auto;
+  width: 614px;
+  background: #fff;
+  margin-top: 15vh;
+  margin-left: calc(50% - 307px);
+}
+.reserveInfo {
+  position: relative;
+  display: grid;
+}
+.reserveInfo h2 {
+  margin: 81px 0px 35px 0;
+  text-align: center;
+  font-size: 46px;
+  font-weight: 900;
+}
+.reserveDesc {
+  margin-bottom: 11px;
+}
+.reserveDesc p {
+  font-size: 18px;
+  line-height: 1.5em;
+  text-align: center;
+}
+.reserveDescTop {
+  font-weight: bold;
+  margin-bottom: 27px;
+}
+.reserveDescCenter {
+  margin-bottom: 54px;
+}
+.reserveDescBottom {
+  margin-bottom: 27px;
+}
+.reserveDescBottom p { 
+  font-size: 15px;
+}
+.reserveBtnWrapper {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  margin-bottom: 43px;
+}
+.reserveBtn {
+  width: 237px;
+  height: 67px;
+  background: rgb(62,204,181);
+  color: #fff;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  cursor: pointer;
+}
